style(markdown): rich meta previews
