+++ packages/markdown/components/Embed/style.scss
@@ -1,0 +1,83 @@
+.markdown-body .embed {
+  $B: '.embed';
+  & {
+    padding: 15px;
+    background: var(--md-code-background, #f6f8fa);
+    color: var(--md-code-text, inherit);
+    border-radius: var(--md-code-radius, var(--markdown-radius, 3px));
+    border-radius: 3px;
+    box-shadow: inset 0 0 0 1px rgba(black, 0);
+    transition: .2s ease;
+    &:hover {
+      box-shadow: inset 0 0 0 1px rgba(black, .1);
+    }
+  }
+  &:empty {
+    display: none;
+  }
+
+  &-media {
+    display: flex;
+    justify-content: center;
+    >:only-child {
+      flex: 1;
+      margin: -15px;
+      border-radius: 0 !important;
+    }
+  }
+
+  &-link {
+    display: flex;
+    align-items: center;
+    text-decoration: none !important;
+  }
+
+  &-body {
+    flex: 1;
+    line-height: 1.3;
+    &, #{$B}-title {
+      font-size: 1.15em;
+      overflow: hidden;
+      white-space: nowrap;
+      text-overflow: ellipsis;
+    }
+    #{$B}-provider {
+      display: block;
+      text-decoration-color: transparent !important;
+      opacity: .88;
+    }
+    &-url {
+      opacity: .75;
+    }
+  }
+
+  &-provider {
+    font-weight: 700;
+    &:before {
+      font-weight: 400;
+      content: 'via ';
+    }
+  }
+
+  &-img {
+    max-width: 5em;
+    margin: 0 .88em 0px 0;
+    padding: 4px;
+    background: white;
+    border: 1px solid rgba(black, .1);
+    border-radius: 3px;
+    transition: inherit;
+  }
+  &:hover #{$B}-img {
+    border: 1px solid rgba(black, .15);
+    box-shadow: -.15em .33em .9em -.3em rgba(black, .15);
+  }
+
+  &-favicon {
+    width: 12px !important;
+    height: 12px !important;
+    margin-top: 1px;
+    margin-right: 6px;
+    margin-bottom: 6px;
+  }
+}
