+++ packages/markdown/processor/parse/magic-block-parser.js
@@ -177,21 +177,19 @@ function tokenize(eat, value) {
       return eat(match)(WrapPinnedBlocks(table, json));
     }
     case 'embed': {
-      json.title = json.title || 'Embed';
+      // json.title = json.title || 'Embed';
       const { title, url, html } = json;
-      json.provider = `@${new URL(url).hostname
+      json.provider = new URL(url).hostname
         .split(/(?:www)?\./)
         .filter(i => i)
-        .join('')}`;
+        .join('.');
       const data = {
+        ...json,
         url,
         html,
         title,
-        provider: json.provider,
-        height: json.height,
-        width: json.width,
-        iframe: json.iframe,
       };
+      console.log(data);
       return eat(match)(
         WrapPinnedBlocks(
           {
@@ -205,8 +203,6 @@ function tokenize(eat, value) {
               },
             ],
             data: {
-              ...data,
-              json,
               hProperties: { ...data, href: url },
               hName: 'rdme-embed',
             },
