+++ packages/markdown/styles/components.scss
@@ -9,3 +9,5 @@
 @import '../components/Callout/style.scss';
 
 @import '../components/Heading/style.scss';
+
+@import '../components/Embed/style.scss';
